+++ src-tauri/src/models/llm_models.rs
@@ -55,7 +55,7 @@ pub struct LLMModel {
     /// Maximum generation length in tokens
     pub max_output_tokens: usize,
     /// Default sampling temperature (0.0 - 2.0)
-    pub temperature_default: f32,
+    pub temperature_default: f64,
     /// Whether this is a builtin model (cannot be deleted)
     pub is_builtin: bool,
     /// Whether this is a reasoning/thinking model (enables thinking output)
@@ -134,7 +134,7 @@ pub struct CreateModelRequest {
     pub max_output_tokens: usize,
     /// Default sampling temperature (0.0 - 2.0, defaults to 0.7)
     #[serde(default = "default_temperature")]
-    pub temperature_default: f32,
+    pub temperature_default: f64,
     /// Whether this is a reasoning/thinking model (defaults to false)
     #[serde(default)]
     pub is_reasoning: bool,
@@ -153,7 +153,7 @@ pub struct CreateModelRequest {
 }
 
 /// Default temperature value for new models.
-fn default_temperature() -> f32 {
+fn default_temperature() -> f64 {
     0.7
 }
 
@@ -242,7 +242,7 @@ pub struct UpdateModelRequest {
     /// New max output tokens (256 - 128,000)
     pub max_output_tokens: Option<usize>,
     /// New default temperature (0.0 - 2.0)
-    pub temperature_default: Option<f32>,
+    pub temperature_default: Option<f64>,
     /// Whether this is a reasoning/thinking model
     pub is_reasoning: Option<bool>,
     /// New price per million input tokens (USD)
